requirements.txt
crewai==0.41.1
crewai-tools==0.4.26
requests==2.31.0
langchain==0.1.20
langchain-openai==0.1.8
python-dotenv==1.0.0
